feat: switch to pyproject.toml
