--- bin/nsgr_submit
@@ -1,10 +1,0 @@
-#!/usr/bin/env python
-import sys
-import os
-import pynsgr.commands as CipresCommands
-
-def main():
-	return CipresCommands.nsgr_submit(sys.argv)
-
-if __name__ == "__main__":
-    sys.exit(main())
