+++ setup.cfg
@@ -1,0 +1,36 @@
+[metadata]
+name = pynsgr
+version = 0.9.5
+author = "Terry Schwartz, Padraig Gleeson"
+author_email = terri@sdsc.edu
+url = https://github.com/OpenSourceBrain/pynsgr
+long_description = file: README.md
+long_description_content_type = text/markdown
+license = MIT
+classifiers=
+    Intended Audience :: Science/Research
+    License :: OSI Approved :: MIT
+    Natural Language :: English
+    Operating System :: OS Independent
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
+    Programming Language :: Python :: 3.11
+    Topic :: Scientific/Engineering
+
+[options]
+install_requires =
+    requests >= 2.5.3
+
+packages = find:
+
+
+[options.packages.find]
+where = .
+include = pynsgr*
+
+[options.entry_points]
+console_scripts =
+    nsgr_submit = pynsgr.commands.nsgr_submit:main
+    nsgr_job = pynsgr.commands.nsgr_job:main
